Convert TestIngestConfig to builder for TestIngestType
